# sdk/log: Fix counting number of dropped attributes of log `Record`