--- a/sdk/log/record_test.go
+++ b/sdk/log/record_test.go
@@ -349,6 +349,8 @@ func TestApplyAttrLimitsDeduplication(t *testing.T) {
 		name        string
 		limit       int
 		input, want log.Value
+
+		droppedAttrs int
 	}{
 		{
 			// No de-duplication
@@ -419,6 +421,7 @@ func TestApplyAttrLimitsDeduplication(t *testing.T) {
 				log.String("g", "GG"),
 				log.String("h", "H"),
 			),
+			droppedAttrs: 10,
 		},
 	}
 
@@ -437,6 +440,8 @@ func TestApplyAttrLimitsDeduplication(t *testing.T) {
 				r.SetAttributes(kv)
 				assertKV(t, r, log.KeyValue{Key: key, Value: tc.want})
 			})
+
+			require.Equal(t, tc.droppedAttrs, r.DroppedAttributes())
 		})
 	}
 }
@@ -446,8 +451,6 @@ func TestApplyAttrLimitsTruncation(t *testing.T) {
 		name        string
 		limit       int
 		input, want log.Value
-
-		expectedDropped int
 	}{
 		{
 			name:  "Empty",
@@ -533,18 +536,6 @@ func TestApplyAttrLimitsTruncation(t *testing.T) {
 				log.Map("7", log.String("a", "")),
 			),
 		},
-		{
-			name:  "MapWithDuplicate",
-			limit: 0,
-			input: log.MapValue(
-				log.Bool("subKey", true),
-				log.Bool("subKey", false),
-			),
-			want: log.MapValue(
-				log.Bool("subKey", false),
-			),
-			expectedDropped: 1,
-		},
 	}
 
 	for _, tc := range testcases {
@@ -562,8 +553,6 @@ func TestApplyAttrLimitsTruncation(t *testing.T) {
 				r.SetAttributes(kv)
 				assertKV(t, r, log.KeyValue{Key: key, Value: tc.want})
 			})
-
-			require.Equal(t, tc.expectedDropped, r.DroppedAttributes())
 		})
 	}
 }