@dmathieu 2024-06-03
Since this is fixing an issue counting the number of dropped attributes, could you add a changelog entry?