@amanakin 2024-06-03
@dmathieu Sure, added changelog entry.