@MrAlias 2024-06-03
> **Before**: `BenchmarkSetAttributes-10 11057461 102.5 ns/op 48 B/op 1 allocs/op`
> **After**: `BenchmarkSetAttributes-10 12738790 86.69 ns/op 48 B/op 1 allocs/op`

Please use [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) to report the statistical significance of these benchmark changes.